refactor: split out header component to isolate render
