+++ packages/core/index.ts
@@ -16,7 +16,7 @@ export { Drawer } from "./components/Drawer";
 
 export { DropZone } from "./components/DropZone";
 export * from "./components/IconButton";
-export * from "./components/Puck";
+export { Puck } from "./components/Puck";
 export * from "./components/Render";
 
 export * from "./lib/migrate";
